migrate profiles to new apps profiles
