+++ lettings/templates/lettings/letting.html
@@ -5,4 +5,4 @@
 <p>{{ address.country_iso_code }}</p>
 <div><a href="{% url 'lettings:lettings_index' %}">Back</a></div>
 <div><a href="{% url 'index' %}">Home</a></div>
-<div><a href="{% url 'profiles_index' %}">Profiles</a></div>
+<div><a href="{% url 'profiles:profiles_index' %}">Profiles</a></div>
