+++ lettings/templates/lettings/lettings_index.html
@@ -14,4 +14,4 @@
     <p>No lettings are available.</p>
 {% endif %}
 <div><a href="{% url 'index' %}">Home</a></div>
-<div><a href="{% url 'profiles_index' %}">Profiles</a></div>
+<div><a href="{% url 'profiles:profiles_index' %}">Profiles</a></div>
