+++ profiles/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+
+from .models import Profile
+
+
+admin.site.register(Profile)
