+++ profiles/templates/profiles/profile.html
@@ -4,6 +4,6 @@
 <p>Last name: {{ profile.user.last_name }}</p>
 <p>Email: {{ profile.user.email }}</p>
 <p>Favorite city: {{ profile.favorite_city }}</p>
-<div><a href="{% url 'profiles_index' %}">Back</a></div>
+<div><a href="{% url 'profiles:profiles_index' %}">Back</a></div>
 <div><a href="{% url 'index' %}">Home</a></div>
 <div><a href="{% url 'lettings:lettings_index' %}">Lettings</a></div>
